fix: refactor names, move methods, delete ambiguous folders
