+++ client/src/manager/RevolutManager.ts
@@ -1,4 +1,4 @@
-import {DateFormat, dateFormatter, formatRevolutDate} from "../util/DateUtils";
+import {DateFormat, dateFormatter, formatRevolutDate} from "../hooks/DateUtils";
 import {ReportManager} from "./ReportManager";
 import {Category} from "../enum/Category";
